Refactored these files so that the imports in the client works. It currently tries importing material modules with the prefix MdMatname, whereas it should be like so: ```javascript import { MatButtonModule } from '@angular/core' ```
Some of the errors that were given when trying to build the client side was:

ERROR in Error: MdButtonModule is not an NgModule
(this error came in a bunch of variations, in fact all material modules, that were referenced).

The client side of the example app, now works by following the original instructions to build.

This commit and PR sets out to resolve issue #8.